Fix: Divided student controller into smaller files
