--- controllers/teacher.js
@@ -1,1 +1,0 @@
-
